+++ src/main/java/com/example/backend/customHashtag/CustomHashtagRepository.java
@@ -1,6 +1,6 @@
-package com.example.backend.video.repository;
+package com.example.backend.customHashtag;
 
-import com.example.backend.video.domain.CustomHashtag;
+import com.example.backend.customHashtag.CustomHashtag;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
